Вынес компонент отображения ингредиента на верхний уровень древа компонентов
